painel

+++ crmZap/app/Http/Controllers/API/AttendantController.php
@@ -89,13 +89,17 @@ class AttendantController extends Controller
     }
 
 
-    public function dellAtendente($id)
+    public function dellAtendente($user_id)
     {
-        $atendente = Atendente::find($id);
-        if ($atendente->delete()) {
-            return response()->json(['type' => 'success']);
+        $atendente = Atendente::where('user_id', $user_id)->first();
+        if ($atendente) {
+            if ($atendente->delete()) {
+                return response()->json(['type' => 'success']);
+            } else {
+                return response()->json(['type' => 'error']);
+            }
         } else {
-            return response()->json(['type' => 'error']);
+            return response()->json(['type' => 'error', 'message' => 'Atendente não encontrado com esse user_id']);
         }
     }
